backend/requirements.txt
# Agentic AI Pharmacy Backend Dependencies
# Pinned compatible versions as of January 2026

# Web Framework
fastapi==0.128.0
uvicorn==0.40.0
starlette==0.50.0

# AI/LLM Stack
openai==2.15.0
langsmith==0.6.3
langchain-openai==1.1.7
langchain-core==1.2.7
langchain-community==0.4.1

# Data Processing & Validation
pydantic==2.12.5
numpy==2.2.6
pandas==2.3.3

# HTTP & Networking
python-dotenv==1.2.1
httpx==0.28.1
requests==2.32.5
websockets==15.0.1

# Utilities
watchfiles==1.1.1
protobuf==4.25.8

# OpenTelemetry (Observability)
opentelemetry-proto==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
